requirements.txt
Authlib==1.6.5
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.0
Flask-Admin==2.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==21.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.0.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dotenv==1.0.1
python-slugify==8.0.4
qrcode==8.0
reportlab==4.2.5
requests==2.32.5
SQLAlchemy==2.0.37
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1

pyTelegramBotAPI==4.29.1
telebot==0.0.5